added the class  that supports averaging.
The main file (screenshoter.py) is now class based on averaging. A cool thing is that in he class you can add you own averaging method as they are dynamically used. but t must only take two inputs (images, out_image_size), and must output two (image(averaged), "explainer about averaging method (short)"). they kust be outputted in that order or else it will thing the image is the explainer. once again make sure it is in a safe directory and not your personal area because it will create and delete files/folders.